+++ bloc_signals/lib/bloc_signals.dart
@@ -11,4 +11,5 @@ export 'src/bloc_signals_base.dart';
 export 'src/concurrency/event_transformers.dart';
 export 'src/concurrency/mutex.dart';
 export 'src/devtools_observer.dart';
+export 'src/devtools_service.dart';
 export 'src/stream_adapter.dart';